+++ spec/punchblock/translator/freeswitch/component/input_spec.rb
@@ -123,6 +123,70 @@ module Punchblock
                 end
               end
 
+              context 'with a builtin grammar' do
+                let(:original_command_opts) { { grammar: { url: 'builtin:dtmf/boolean' } } }
+
+                before do
+                  subject.execute
+                  expected_event
+                  send_dtmf 1
+                end
+
+                let :expected_nlsml do
+                  RubySpeech::NLSML.draw do
+                    interpretation confidence: 1 do
+                      instance "true"
+                      input "1", mode: :dtmf
+                    end
+                  end
+                end
+
+                let :expected_event do
+                  Punchblock::Component::Input::Complete::Match.new nlsml: expected_nlsml
+                end
+
+                it "should use RubySpeech builtin grammar" do
+                  reason.should be == expected_event
+                end
+              end
+
+              context 'with a parameterized builtin grammar' do
+                let(:original_command_opts) { { grammar: { url: 'builtin:dtmf/boolean?n=3;y=4' } } }
+
+                before do
+                  subject.execute
+                  expected_event
+                  send_dtmf 4
+                end
+
+                let :expected_nlsml do
+                  RubySpeech::NLSML.draw do
+                    interpretation confidence: 1 do
+                      instance "true"
+                      input "4", mode: :dtmf
+                    end
+                  end
+                end
+
+                let :expected_event do
+                  Punchblock::Component::Input::Complete::Match.new nlsml: expected_nlsml
+                end
+
+                it "should use RubySpeech builtin grammar" do
+                  reason.should be == expected_event
+                end
+              end
+
+              context 'with a bad builtin grammar name' do
+                let(:original_command_opts) { { grammar: { url: 'builtin:dtmf/foobar' } } }
+
+                it "should return an error and not execute any actions" do
+                  subject.execute
+                  error = ProtocolError.new.setup 'option error', 'foobar is an invalid builtin grammar'
+                  original_command.response(0.1).should be == error
+                end
+              end
+
               context 'with multiple grammars' do
                 let(:original_command_opts) { { :grammars => [{:value => grammar}, {:value => grammar}] } }
                 it "should return an error and not execute any actions" do