+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/adhearsion/punchblock)
+  * Feature: Support RubySpeech builtin grammars on Asterisk and FreeSWITCH
 
 # [v2.0.2](https://github.com/adhearsion/punchblock/compare/v2.0.1...v2.0.2) - [2013-10-17](https://rubygems.org/gems/punchblock/versions/2.0.2)
   * Bugfix: Reject commands against components which have finished on Asterisk, and garbage collect them